+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -59,5 +59,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-07-21T05:52:38.303Z")
+    created: Optional[datetime] = parse_datetime("2025-07-21T08:30:49.315Z")
     r"""The date and time the resource was created"""