+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -93,5 +93,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-07-21T05:52:38.303Z")
+    created: Optional[datetime] = parse_datetime("2025-07-21T08:30:49.315Z")
     r"""The date and time the resource was created"""