+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K0NRN9J2XZ6FJ6QE2MCDTVP9"
+        "01K0P1PY79S1Y0XN0W3JDXXW0C"
     )
     r"""The id of the resource"""