+++ packages/orq-rc/RELEASES.md
@@ -2029,3 +2029,13 @@ Based on:
 - [python v3.10.0-rc.11] packages/orq-rc
 ### Releases
 - [PyPI v3.10.0-rc.11] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.11 - packages/orq-rc
+
+## 2025-07-21 08:32:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.583.0 (2.658.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.10.0-rc.12] packages/orq-rc
+### Releases
+- [PyPI v3.10.0-rc.12] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.12 - packages/orq-rc